refactor: extract CompoundNode.tsx from overlays
